+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## 0.10.0 (tbd)
 
 * Introduce compatibility check for the Piral CLI (#94)
-* Added new optional plugin `piral-litel` for Lit-Element (#85)
-* Added new optional plugin `piral-aurelia` for Aurelia (#80)
 * Added new optional plugin `piral-mithril` for Mithril.js (#79)
+* Added new optional plugin `piral-aurelia` for Aurelia (#80)
+* Added new optional plugin `piral-litel` for Lit-Element (#85)
+* Added new optional plugin `piral-svelte` for Svelte (#97)
+* Added new optional plugin `piral-elm` for Elm (#118)
 * Replaced `react-arbiter` with `piral-base` (#109)
 
 ## 0.9.3 (December 17, 2019)